@@ -15,23 +15,24 @@
     "core-js": "^2.6.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/animations": "^7.2.15",
-    "@angular/cli": "^7.3.8",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/platform-server": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/animations": "^8.0.1",
+    "@angular/cdk": "^8.0.1",
+    "@angular/cli": "^8.0.3",
+    "@angular/common": "^8.0.1",
+    "@angular/compiler": "^8.0.1",
+    "@angular/compiler-cli": "^8.0.1",
+    "@angular/core": "^8.0.1",
+    "@angular/forms": "^8.0.1",
+    "@angular/language-service": "^8.0.1",
+    "@angular/platform-browser": "^8.0.1",
+    "@angular/platform-browser-dynamic": "^8.0.1",
+    "@angular/platform-server": "^8.0.1",
+    "@angular/router": "^8.0.1",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^6.14.2",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "^4.0.1",
@@ -41,14 +42,14 @@
     "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "primeicons": "^1.0.0",
-    "primeng": "^7.0.3",
+    "primeng": "^8.0.0",
     "protractor": "^5.4.2",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.2",
     "ts-node": "~3.0.4",
     "tslint": "^5.12.0",
-    "typescript": "^3.1.6",
+    "typescript": "^3.4.5",
     "webpack": "^4.29.0",
     "webpack-dev-server": "^3.1.14",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   }
 }
